requirements.txt
discord.py
requests
spotipy
yt_dlp
pynacl
python-dotenv
fastapi
uvicorn
motor
pandas
scikit-learn